dev_requirements.txt
tox
coverage

django<=4
django-appconf
django-cms==4.1rc4
djangocms-text-ckeditor